Infinite scrolling done!

+++ minimalist-game-framework-core/Game/Globals.cs
@@ -5,6 +5,6 @@ using System.Text;
 public static class Globals
 {
     public static readonly string TITLE = "Jetpack Joyride";
-    public static readonly int WIDTH = 1000;
-    public static readonly int HEIGHT = 500;
+    public static readonly float WIDTH = 1000;
+    public static readonly float HEIGHT = 500;
 }
